chore(nlu): standardize error throwing in nlu components
