+++ src/bp/nlu-core/slots/typings.ts
@@ -1,0 +1,13 @@
+import { Tag } from 'nlu-core/typings'
+
+export interface TagResult {
+  tag: Tag | string
+  name: string
+  probability: number
+}
+
+export interface IntentSlotFeatures {
+  name: string
+  vocab: string[]
+  slot_entities: string[]
+}
